requirements.txt
# ============================================
# Clubhouse-Podcast-Automation Dependencies
# ============================================

# -----------------------------
# CORE (Required for all modes)
# -----------------------------
click>=8.0.0              # CLI framework
pyyaml>=6.0               # Configuration parsing
python-dotenv>=1.0.0      # Environment variable loading
requests>=2.28.0          # HTTP requests
tqdm>=4.65.0              # Progress bars

# -----------------------------
# LOCAL + CLOUD: Audio/Video Processing
# -----------------------------
moviepy>=1.0.3            # Video editing and composition
librosa>=0.10.0           # Audio analysis (waveform extraction)
numpy>=1.24.0             # Numerical operations
Pillow>=10.0.0            # Image processing
soundfile>=0.12.0         # Audio file I/O (required by librosa)

# -----------------------------
# LOCAL + CLOUD: AI APIs (Gemini)
# -----------------------------
google-generativeai>=0.3.0  # Gemini API for transcription & summarization

# -----------------------------
# LOCAL + CLOUD: YouTube Upload
# -----------------------------
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# -----------------------------
# CLOUD ONLY: GCP Storage
# (Not needed for local deployment)
# -----------------------------
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0

# -----------------------------
# DEVELOPMENT (Optional)
# -----------------------------
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
